Visualizer: Cleaned up.

+++ src/inet/visualizer/base/PathCanvasVisualizerBase.cc
@@ -93,7 +93,7 @@ void PathCanvasVisualizerBase::refreshDisplay() const
         auto pathCanvasVisualization = static_cast<const PathCanvasVisualization *>(pathVisualization);
         auto moduleIds = pathCanvasVisualization->moduleIds;
         std::vector<LineSegment> segments;
-        for (int index = 1; index < moduleIds.size(); index++) {
+        for (int index = 1; index < (int)moduleIds.size(); index++) {
             auto fromModuleId = moduleIds[index - 1];
             auto toModuleId = moduleIds[index];
             auto fromModule = simulation->getModule(fromModuleId);
@@ -104,7 +104,7 @@ void PathCanvasVisualizerBase::refreshDisplay() const
             segments.push_back(LineSegment(fromPosition + shift, toPosition + shift));
         }
         std::vector<cFigure::Point> points;
-        for (int index = 0; index < segments.size(); index++) {
+        for (int index = 0; index < (int)segments.size(); index++) {
             if (index == 0)
                 points.push_back(canvasProjection->computeCanvasPoint(segments[index].getPoint1()));
             if (index > 0) {
@@ -131,7 +131,7 @@ void PathCanvasVisualizerBase::refreshDisplay() const
                     points.push_back(canvasProjection->computeCanvasPoint(segment2.getPoint1()));
                 }
             }
-            if (index == segments.size() - 1)
+            if (index == (int)segments.size() - 1)
                 points.push_back(canvasProjection->computeCanvasPoint(segments[index].getPoint2()));
         }
         pathCanvasVisualization->figure->setPoints(points);
